+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`0.1.0...main`][0.1.0...main].
 ### Changed
 
 * Renamed `InvalidDefinition::fromClassNameAndException()` to `InvalidDefinition::throwsExceptionDuringInstantiation()` ([#300]), by [@localheinz]
+* Renamed `Number` to `Count` ([#309]), by [@localheinz]
 
 ### Fixed
 
@@ -146,5 +147,6 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [#300]: https://github.com/ergebnis/factory-bot/pull/300
 [#301]: https://github.com/ergebnis/factory-bot/pull/301
 [#302]: https://github.com/ergebnis/factory-bot/pull/302
+[#309]: https://github.com/ergebnis/factory-bot/pull/309
 
 [@localheinz]: https://github.com/localheinz